64 stuff

+++ src/bootloader/stage2/stdio.h
@@ -5,4 +5,4 @@ void clrscr();
 void putc(char c);
 void puts(const char* str);
 void printf(const char* fmt, ...);
-void print_buffer(const char* msg, const void* buffer, uint32_t count);
+void print_buffer(const char* msg, const void* buffer, uint64_t count);
